government_up_yet/requirements.txt
APScheduler==3.5.3
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
more-itertools==5.0.0
pluggy==0.8.0
psycopg2-binary==2.7.6.1
py==1.7.0
pytest==4.0.2
python-dateutil==2.7.5
pytz==2018.7
requests==2.21.0
six==1.12.0
SQLAlchemy==1.2.15
tzlocal==1.5.1
urllib3==1.24.1
Werkzeug==0.14.1
